requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-nested-admin==3.3.2
django-phonenumber-field==4.0.0
django-widget-tweaks==1.4.8
idna==2.10
phonenumberslite==8.12.7
Pillow==7.2.0
python-monkey-business==1.0.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
